pth threads enabled

+++ src/lib/libzpool/include/sys/zfs_context.h
@@ -443,7 +443,7 @@ extern void delay(clock_t ticks);
 #define	minclsyspri	60
 #define	maxclsyspri	99
 
-#define	CPU_SEQID	(thr_self() & (max_ncpus - 1))
+#define	CPU_SEQID	((uint32_t)thr_self() & (max_ncpus - 1))
 
 #define	kcred		NULL
 #define	CRED()		NULL
